Discover mujoco using environment variables
Adds the option to locate mjpro150 and mjkey.txt using environment
variables.

This is extremely useful when using mujoco_py in a cluster environment
or with Docker.

* MUJOCO_PY_MUJOCO_PATH
Provides an alternative path to the directory that is usually
~/.mujoco.
If this environment variable is set and non-empty, it is used as the
directory prefix for mjpro150 and mjkey.txt, replacing ~/.mujoco.

* MUJOCO_LICENSE_PATH
Provides an alternative path to the MuJoCo Pro key.
If set and non-empty, this environment variable overrides the path to
the file containing the MuJoCo Pro license key, including any
alternative path induced by MUJOCO_PY_MUJOCO_PATH.

* MUJOCO_PY_MJPRO_PATH
Provides and alternative path to the mjpro150 binaries.
If set and non-empty, this environment variable overrides the path to
the mjpro150 binaries, including any alternative path induced by
MUJOCO_PY_MUJOCO_PATH.