Add MacPort's gcc to supported compilers.
I've got stuck during installation, because I haven't noticed that there is a fixed list of supported gcc versions, which is apparently defined just based on brew versions. I'm proposing edit to add MacPorts gcc version as well, I haven't test the `gcc-mp-9`, but all the others seems to work. Other approach could be to actually use system gcc and parse output of `gcc --version` to know if the version is actually supported or not.